# swagger.yaml
openapi: 3.0.3
info:
  title: eKalakaar API
  description: >-
    Explore the eKalakaar API documentation, a comprehensive guide to the features and functionalities of the eKalakaar platform's APIs. From user registration to authentication and password reset, this document covers the various endpoints that power eKalakaar's creative ecosystem.

    ## Dummy Users -
      ### Patron:
        email: william.wilson@example.com
        password: william123

        email: "emma.davis@example.com
        password: emma456

        email: liam.rodriguez@example.com
        password: liam789

        email: ava.hernandez@example.com
        password: ava000

        email: noah.gonzalez@example.com
        password: noah777
      ### Artist:
        email: emily.anderson@example.com
        password: emily123

        email: michael.johnson@example.com
        password: michael456

        email: sophia.garcia@example.com
        password: sophia789

        email: david.martinez@example.com
        password: david000

        email: olivia.brown@example.com
        password: olivia777
    ### User Avatar/Profile image Upload Info:

    Default user avatar server directory - /images/image-name.ext.

      where, ext can be any image (eg. png, jpg, jpeg etc).

      To retrieve user avatar pass the user.avatar into the following url - 

      eg. user.avatar have value "user-avatar.png"

      `https://api.ekalakaar.com/images/user-avatar.png` with GET request
  version: 1.0.0
...
servers:
...
paths:
...
  /patrons/artists:
    get:
      tags:
        - Patron
      summary: Get top artists profiles
      description: >-
        Retrieves a list of top artists who have previously applied and been hired for opportunities posted by the logged-in patron.
      responses:
        200:
          description: Artist profiles fetched successfully
        400:
          description: Bad request
        404:
          description: Not found
        401:
          description: Unauthorized
        500:
          description: Internal server error
  /patrons/artists/{id}:
    get:
      tags:
        - Patron
      summary: Get profile info of single artist
      description: >-
        Retrieves profile info of a single artist by its id.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: Artist profile fetched successfully
        400:
          description: Bad request
        404:
          description: Not found
        401:
          description: Unauthorized
        500:
          description: Internal server error